Integración de estados, clases y mejoras backend

+++ aeropuerto/src/main/java/aeropuerto/repository/BoletoRepository.java
@@ -3,6 +3,6 @@ package aeropuerto.repository;
 import aeropuerto.model.Boleto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BoletoRepository extends JpaRepository<Boleto,Integer> {
+public interface BoletoRepository extends JpaRepository<Boleto,Long> {
     boolean existsByAsiento_IdAndVuelo_Id(Long asientoId, Long vueloId);
 }
